Rename "Reflection" to "Frame"
I got the domain model a bit gummed up between the last commit and now.
I want "Reflections" to model how people see a user, so they have to
line up with a set of answers instead of a set of questions. The "Frame"
is where these different composite images of a user come together.